+++ webcomponents/word-cloud/src/components/word-cloud/word-cloud.tsx
@@ -57,7 +57,7 @@ export class DeckdeckgoWordCloud implements DeckdeckgoComponent {
 
   private svgRef!: SVGGElement;
 
-  private colors: string[] = Array.from({length: 5}, (_v, _i) => Math.floor(Math.random() * 16777215).toString(16));
+  private colors: string[] | undefined;
 
   private width: number;
   private height: number;
@@ -189,13 +189,17 @@ export class DeckdeckgoWordCloud implements DeckdeckgoComponent {
       return;
     }
 
+    if (!this.colors || this.colors.length !== words.length) {
+      this.colors = Array.from({length: words.length}, (_v, _i) => Math.floor(Math.random() * 16777215).toString(16));
+    }
+
     const layout = cloud()
       .size([this.width, this.height])
       .words(
         words.map((d: string, index: number) => ({
           text: d,
           size: 10 + Math.random() * 110,
-          color: `var(--deckgo-word-count-fill-color-${index}, ${this.getRandomColor()})`,
+          color: `var(--deckgo-word-count-fill-color-${index}, #${this.colors?.[index]})`,
         }))
       )
       .rotate(() => ~~(Math.random() * 2) * 90)
@@ -207,11 +211,6 @@ export class DeckdeckgoWordCloud implements DeckdeckgoComponent {
       });
 
     layout.start();
-  }
-
-  private getRandomColor(): string {
-    const index: number = Math.floor(Math.random() * this.colors.length);
-    return `#${this.colors[index]}`;
   }
 
   private async applyChanges() {