+++ webcomponents/word-cloud/src/components/word-cloud/word-cloud.tsx
@@ -62,6 +62,8 @@ export class DeckdeckgoWordCloud implements DeckdeckgoComponent {
   private width: number;
   private height: number;
 
+  private editWords: string;
+
   async componentDidLoad() {
     this.initWindowResize();
 
@@ -189,6 +191,13 @@ export class DeckdeckgoWordCloud implements DeckdeckgoComponent {
       return;
     }
 
+    const sizeMatch: boolean = this.svgRef.getAttribute('width') === `${this.width}` && this.svgRef.getAttribute('height') === `${this.height}`;
+
+    if (words.join('') === this.editWords && sizeMatch) {
+      // We do not repaint if the list of words did not change
+      return;
+    }
+
     if (!this.colors || this.colors.length !== words.length) {
       this.colors = Array.from({length: words.length}, (_v, _i) => Math.floor(Math.random() * 16777215).toString(16));
     }
@@ -211,6 +220,8 @@ export class DeckdeckgoWordCloud implements DeckdeckgoComponent {
       });
 
     layout.start();
+
+    this.editWords = words?.join('');
   }
 
   private async applyChanges() {