+++ webcomponents/word-cloud/CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.1.0 (2021-09-02)
 
+### Features
+
+- preserve color if content does not change
+
 ### Build
 
 - bump dependencies